@@ -7,25 +7,46 @@
         "php": ">=8.0.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "doctrine/annotations": "^1.13",
+        "doctrine/annotations": "^1.0",
         "doctrine/doctrine-bundle": "^2.5",
         "doctrine/doctrine-migrations-bundle": "^3.2",
         "doctrine/orm": "^2.11",
         "friendsofsymfony/ckeditor-bundle": "^2.4",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "phpstan/phpdoc-parser": "^1.2",
+        "sensio/framework-extra-bundle": "^6.1",
         "sonata-project/admin-bundle": "^4.9",
+        "symfony/asset": "6.0.*",
         "symfony/console": "6.0.*",
+        "symfony/doctrine-messenger": "6.0.*",
         "symfony/dotenv": "6.0.*",
+        "symfony/expression-language": "6.0.*",
         "symfony/flex": "^2",
+        "symfony/form": "6.0.*",
         "symfony/framework-bundle": "6.0.*",
-        "symfony/maker-bundle": "^1.38",
+        "symfony/http-client": "6.0.*",
+        "symfony/intl": "6.0.*",
+        "symfony/mailer": "6.0.*",
+        "symfony/mime": "6.0.*",
+        "symfony/monolog-bundle": "^3.0",
+        "symfony/notifier": "6.0.*",
+        "symfony/process": "6.0.*",
+        "symfony/property-access": "6.0.*",
+        "symfony/property-info": "6.0.*",
         "symfony/proxy-manager-bridge": "6.0.*",
         "symfony/runtime": "6.0.*",
         "symfony/security-bundle": "6.0.*",
+        "symfony/serializer": "6.0.*",
+        "symfony/string": "6.0.*",
+        "symfony/translation": "6.0.*",
         "symfony/twig-bundle": "6.0.*",
-        "symfony/webpack-encore-bundle": "^1.14",
+        "symfony/validator": "6.0.*",
+        "symfony/web-link": "6.0.*",
+        "symfony/webapp-meta": "^1.0",
+        "symfony/webpack-encore-bundle": "^1.12",
         "symfony/yaml": "6.0.*",
-        "twig/extra-bundle": "^3.3",
-        "twig/twig": "^3.3"
+        "twig/extra-bundle": "^2.12|^3.0",
+        "twig/twig": "^2.12|^3.0"
     },
     "config": {
         "allow-plugins": {
@@ -64,12 +85,10 @@
             "ckeditor:install": "symfony-cmd"
         },
         "post-install-cmd": [
+            "@auto-scripts"
         ],
         "post-update-cmd": [
             "@auto-scripts"
-        ],
-        "compile": [
-            "php bin/console doctrine:migrations:migrate"
         ]
     },
     "conflict": {
@@ -80,5 +99,11 @@
             "allow-contrib": false,
             "require": "6.0.*"
         }
+    },
+    "require-dev": {
+        "symfony/debug-bundle": "6.0.*",
+        "symfony/maker-bundle": "^1.0",
+        "symfony/stopwatch": "6.0.*",
+        "symfony/web-profiler-bundle": "6.0.*"
     }
 }
